chore(login) small refactoring for slideshow
