+++ app/src/entourage/java/social/entourage/android/authentification/login/slideshow/pages/LoginSlideshowPage3Fragment.java
@@ -9,23 +9,14 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import social.entourage.android.R;
+import social.entourage.android.authentication.login.slideshow.LoginSlideshowPageFragment;
 
 /**
- * Login slideshow page 3 {@link Fragment}.
+ * Login slideshow page 3 {@link LoginSlideshowPageFragment}.
  */
-public class LoginSlideshowPage3Fragment extends Fragment {
-
-
+public class LoginSlideshowPage3Fragment  extends LoginSlideshowPageFragment {
     public LoginSlideshowPage3Fragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login_slideshow_page3, container, false);
+        super(R.layout.fragment_login_slideshow_page3);
     }
 
 }
