+++ app/src/entourage/java/social/entourage/android/authentification/login/slideshow/pages/LoginSlideshowPage4Fragment.java
@@ -1,31 +1,17 @@
 package social.entourage.android.authentification.login.slideshow.pages;
 
 
-import android.os.Bundle;
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
 
 import social.entourage.android.R;
+import social.entourage.android.authentication.login.slideshow.LoginSlideshowPageFragment;
 
 /**
  * Login slideshow page 4 {@link Fragment}.
  */
-public class LoginSlideshowPage4Fragment extends Fragment {
-
-
+public class LoginSlideshowPage4Fragment extends LoginSlideshowPageFragment {
     public LoginSlideshowPage4Fragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login_slideshow_page4, container, false);
+        super(R.layout.fragment_login_slideshow_page4);
     }
 
 }
