+++ app/src/main/java/social/entourage/android/authentication/login/slideshow/LoginSlideshowPageFragment.java
@@ -1,23 +1,24 @@
-package social.entourage.android.authentification.login.slideshow.pages;
+package social.entourage.android.authentication.login.slideshow;
 
 
 import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import social.entourage.android.R;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
 /**
  * Login Slideshow page 1 {@link Fragment}.
  */
-public class LoginSlideshowPage1Fragment extends Fragment {
+public class LoginSlideshowPageFragment extends Fragment {
 
+    private int layout;
 
-    public LoginSlideshowPage1Fragment() {
-        // Required empty public constructor
+    public LoginSlideshowPageFragment(int page_layout) {
+        super();
+        layout = page_layout;
     }
