+++ app/src/pfp/java/social/entourage/android/authentification/login/slideshow/pages/LoginSlideshowPage1Fragment.java
@@ -1,31 +1,15 @@
 package social.entourage.android.authentification.login.slideshow.pages;
 
 
-import android.os.Bundle;
-import androidx.annotation.NonNull;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
 import social.entourage.android.R;
+import social.entourage.android.authentication.login.slideshow.LoginSlideshowPageFragment;
 
 /**
- * Login slideshow page 1 {@link Fragment}.
+ * Login slideshow page 1 {@link LoginSlideshowPageFragment}.
  */
-public class LoginSlideshowPage1Fragment extends Fragment {
-
-
+public class LoginSlideshowPage1Fragment  extends LoginSlideshowPageFragment {
     public LoginSlideshowPage1Fragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_login_slideshow_page1, container, false);
+        super(R.layout.fragment_login_slideshow_page1);
     }
 
 }
